--- test/test_dockergen/.gitignore
@@ -1,1 +1,0 @@
-nginx.tmpl
