+++ test/test_dockergen_v2/test_dockergen_v2.yml
@@ -14,7 +14,7 @@ services:
       - nginx
     volumes:
       - /var/run/docker.sock:/tmp/docker.sock:ro
-      - ./nginx.tmpl:/etc/docker-gen/templates/nginx.tmpl
+      - ../nginx.tmpl:/etc/docker-gen/templates/nginx.tmpl
 
   web:
     image: web
