+++ test/test_dockergen_v3/test_dockergen_v3.py
@@ -1,6 +1,5 @@
-import os
+from time import sleep
 import docker
-import logging
 import pytest
 
 
@@ -27,26 +26,9 @@ pytestmark = pytest.mark.skipif(
     reason="Docker compose syntax v3 requires docker engine v1.13 or later (got %s)" % raw_version)
 
 
-@pytest.yield_fixture(scope="module")
-def nginx_tmpl():
-    """
-    pytest fixture which extracts the the nginx config template from
-    the jwilder/nginx-proxy:test image
-    """
-    script_dir = os.path.dirname(__file__)
-    logging.info("extracting nginx.tmpl from jwilder/nginx-proxy:test")
-    docker_client = docker.from_env()
-    print(docker_client.containers.run(
-        image='jwilder/nginx-proxy:test',
-        remove=True,
-        volumes=['{current_dir}:{current_dir}'.format(current_dir=script_dir)],
-        entrypoint='sh',
-        command='-xc "cp /app/nginx.tmpl {current_dir} && chmod 777 {current_dir}/nginx.tmpl"'.format(
-            current_dir=script_dir),
-        stderr=True))
-    yield
-    logging.info("removing nginx.tmpl")
-    os.remove(os.path.join(script_dir, "nginx.tmpl"))
+def test_nginx_is_running(nginx_tmpl, docker_compose):
+    sleep(3)
+    assert docker_compose.containers.get("nginx").status == "running"
 
 
 def test_unknown_virtual_host_is_503(nginx_tmpl, docker_compose, nginxproxy):
@@ -59,8 +41,3 @@ def test_forwards_to_whoami(nginx_tmpl, docker_compose, nginxproxy):
     assert r.status_code == 200
     whoami_container = docker_compose.containers.get("whoami")
     assert r.text == "I'm %s\n" % whoami_container.id[:12]
-
-
-if __name__ == '__main__':
-    import doctest
-    doctest.testmod()
