+++ test/test_dockergen_v3/test_dockergen_v3.yml
@@ -1,4 +1,5 @@
 version: '3'
+
 services:
   nginx:
     image: nginx
@@ -11,7 +12,7 @@ services:
     command: -notify-sighup nginx -watch /etc/docker-gen/templates/nginx.tmpl /etc/nginx/conf.d/default.conf
     volumes:
       - /var/run/docker.sock:/tmp/docker.sock:ro
-      - ./nginx.tmpl:/etc/docker-gen/templates/nginx.tmpl
+      - ../nginx.tmpl:/etc/docker-gen/templates/nginx.tmpl
       - nginx_conf:/etc/nginx/conf.d
 
   web:
